@@ -3,13 +3,13 @@
   "private": true,
   "version": "0.6.1",
   "engines": {
-    "node": "16.x"
+    "node": "18.x"
   },
   "main": "./src/index.js",
   "type": "module",
   "scripts": {
-    "iterate": "concurrently \"node build.js --watch\" \"nodemon\"",
-    "build": "rimraf ./build && node build.js --build",
+    "build": "rimraf ./build && tsx ./build.ts --build",
+    "iterate": "concurrently \"tsx ./build.ts --watch\" \"nodemon\"",
     "start": "NODE_ENV=production node build/index.js",
     "type-check": "tsc --noEmit",
     "lint": "eslint \"./src/**/*.{ts,}\" --max-warnings 0",
@@ -23,7 +23,7 @@
     "express": "4.18.2",
     "express-ws": "5.0.2",
     "http-proxy": "^1.18.1",
-    "networked-dom-server": "0.6.0",
+    "networked-dom-server": "0.7.0",
     "ws": "8.13.0"
   },
   "devDependencies": {
@@ -31,7 +31,7 @@
     "@types/express": "^4.17.17",
     "@types/express-ws": "^3.0.1",
     "@types/http-proxy": "^1.17.11",
-    "@types/node": "^20.3.0",
-    "nodemon": "^2.0.22"
+    "@types/node": "^20.5.7",
+    "nodemon": "^3.0.1"
   }
 }
