@@ -1,8 +1,9 @@
 {
   "name": "3d-web-experience",
   "private": true,
+  "type": "module",
   "engines": {
-    "node": "16.x"
+    "node": "18.x"
   },
   "scripts": {
     "preinstall": "node check-node-version.js",
@@ -22,24 +23,29 @@
     "example/*"
   ],
   "devDependencies": {
-    "@types/jest": "29.5.2",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "concurrently": "^8.1.0",
-    "esbuild": "0.17.19",
+    "@types/jest": "29.5.4",
+    "@typescript-eslint/eslint-plugin": "^6.5.0",
+    "@typescript-eslint/parser": "^6.5.0",
+    "concurrently": "^8.2.1",
+    "esbuild": "0.19.2",
     "esbuild-jest": "0.5.0",
     "esbuild-plugin-copy": "2.1.1",
-    "esbuild-plugin-d.ts": "^1.1.0",
-    "eslint": "^8.42.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.2.1",
-    "jest": "29.5.0",
-    "jest-environment-jsdom": "29.5.0",
-    "lerna": "6.6.2",
-    "prettier": "^2.8.8",
+    "jju": "^1.4.0",
+    "@types/jju": "^1.4.2",
+    "tmp": "^0.2.1",
+    "@types/tmp": "^0.2.3",
+    "eslint": "^8.48.0",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "jest": "29.6.4",
+    "jest-environment-jsdom": "29.6.4",
+    "lerna": "7.2.0",
+    "prettier": "^3.0.2",
     "rimraf": "^5.0.1",
-    "typescript": "^5.0.2"
+    "ts-node": "^10.9.1",
+    "tsx": "3.12.7",
+    "typescript": "^5.2.2"
   }
 }
